+++ src/main/java/com/devs/ecommerce/model/Order.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -28,8 +29,8 @@ public class Order {
     private String address;
     private String town;
     private String phoneNumber;
-    private double deliveryTax;
-    private double totalPrice;
+    private BigDecimal deliveryTax;
+    private BigDecimal totalPrice;
     private OrderStatus status;
     public enum OrderStatus {
         PENDING, PREPARING, SHIPPED, DELIVERING, DELIVERED, CANCELED
